fixed issue of file being reset after server restart

+++ src/main/java/com/dufji/stark/database/impl/StarkFileDatabase.java
@@ -2,10 +2,7 @@ package com.dufji.stark.database.impl;
 
 import com.dufji.stark.Stark;
 import com.dufji.stark.database.StarkDatabase;
-import com.dufji.stark.model.StarkPlayer;
 import com.google.gson.*;
-import org.bukkit.Bukkit;
-import org.bukkit.OfflinePlayer;
 
 import java.io.File;
 import java.io.FileReader;
@@ -24,13 +21,18 @@ public class StarkFileDatabase extends StarkDatabase {
             File dataFolder = new File(Stark.getInstance().getDataFolder().getPath(), "data");
             if (!dataFolder.exists()) dataFolder.mkdirs();
             File dataFile = new File(dataFolder, "data.json");
-            dataFile.createNewFile();
-            JsonObject jsonObject = new JsonObject();
-            FileWriter fileWriter = new FileWriter(dataFile);
-            fileWriter.write(gson.toJson(jsonObject));
-            fileWriter.flush();
-            fileWriter.close();
-            this.jsonObject = jsonObject;
+            if(!dataFile.exists()) {
+                dataFile.createNewFile();
+                FileWriter fileWriter = new FileWriter(dataFile);
+                JsonObject jsonObject = new JsonObject();
+                fileWriter.write(gson.toJson(jsonObject));
+                fileWriter.flush();
+                fileWriter.close();
+                this.jsonObject = jsonObject;
+            } else {
+                this.jsonObject = new JsonParser().parse(new FileReader(dataFile)).getAsJsonObject();
+
+            }
         } catch (Exception exception) {
             Stark.getInstance().getLogger().log(Level.SEVERE, "An error occurred while initializing the file database.");
         }
@@ -39,9 +41,9 @@ public class StarkFileDatabase extends StarkDatabase {
     private void saveJsonObject() {
         try {
             File dataFolder = new File(Stark.getInstance().getDataFolder().getPath(), "data");
+            if (!dataFolder.exists()) dataFolder.mkdirs();
             File dataFile = new File(dataFolder, "data.json");
-            dataFile.delete();
-            dataFile.createNewFile();
+            if (!dataFile.exists()) dataFile.createNewFile();
             FileWriter fileWriter = new FileWriter(dataFile);
             fileWriter.write(gson.toJson(this.jsonObject));
             fileWriter.flush();
@@ -73,7 +75,6 @@ public class StarkFileDatabase extends StarkDatabase {
         return jsonObject.get(uuid.toString()).getAsFloat();
     }
 
-    //TODO: Fix the issue of the balance resetting to 0 after a server restart
     @Override
     public void setBalance(UUID uuid, float balance) {
         refreshJsonObject();
@@ -89,7 +90,6 @@ public class StarkFileDatabase extends StarkDatabase {
 
 
 
-
     @Override
     public Integer getBalTopPosition(UUID uuid) {
         refreshJsonObject();
@@ -100,22 +100,17 @@ public class StarkFileDatabase extends StarkDatabase {
             return 0;
         }
 
+        Map<UUID, Float> map = new HashMap<>();
+        for (Map.Entry<String, JsonElement> entry : jsonObject.entrySet()) {
+            map.put(UUID.fromString(entry.getKey()), entry.getValue().getAsFloat());
+        }
 
-        OfflinePlayer[] allPlayers = Bukkit.getOfflinePlayers();
+        LinkedHashMap<UUID, Float> sortedMap = map.entrySet().stream()
+                .sorted(Map.Entry.comparingByValue(Comparator.reverseOrder()))
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (e1, e2) -> e1, LinkedHashMap::new));
 
-        List<StarkPlayer> sortedPlayers = Arrays.stream(allPlayers)
-                .map(player -> new StarkPlayer(player.getUniqueId()))
-                .sorted(Comparator.comparing(StarkPlayer::getBalance).reversed())
-                .collect(Collectors.toList());
-
-
-
-        List<StarkPlayer> topPlayers = sortedPlayers.stream()
-                .limit(jsonObject.entrySet().size())
-                .collect(Collectors.toList());
-
-
-        return topPlayers.indexOf(new StarkPlayer(uuid));
+        List<UUID> list = new ArrayList<>(sortedMap.keySet());
+        return list.indexOf(uuid) + 1;
     }
 
 }
